src/requirements.txt
requests
pandas
numpy
gymnasium
stable_baselines3
optuna
torch
pyyaml
pandas_datareader
talib
